pyproject.toml
[tool.poetry]
name = "drmhbot"
version = "0.1.0"
description = ""
authors = ["Julia <juliariety@gmail.com>"]
license = "Proprietary"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.27.1"
beautifulsoup4 = "^4.10.0"
python-telegram-bot = "^13.11"
redis = "^4.1.4"
hy = "^0.20.0"
# Work-around for "error in funcparserlib setup command: use_2to3 is invalid"
funcparserlib = "1.0.0a0"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
isort = "^5.10.1"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
